refactor: context 의 값을 컴포넌트화
